+++ services/horizon/internal/simplepath/order_book_test.go
@@ -78,39 +78,6 @@ func TestOrderBook_BadCost(t *testing.T) {
 	}
 }
 
-func TestConvertToBuyingUnits(t *testing.T) {
-	testCases := []struct {
-		sellingOfferAmount int64
-		sellingUnitsNeeded int64
-		pricen             int64
-		priced             int64
-		wantBuyingUnits    int64
-		wantSellingUnits   int64
-	}{
-		{7, 2, 3, 7, 1, 2},
-		{math.MaxInt64, 2, 3, 7, 1, 2},
-		{20, 20, 1, 4, 5, 20},
-		{20, 100, 1, 4, 5, 20},
-		{20, 20, 7, 11, 13, 19},
-		{20, 20, 11, 7, 32, 20},
-		{20, 100, 7, 11, 13, 19},
-		{20, 100, 11, 7, 32, 20},
-		{1, 0, 3, 7, 0, 0},
-		{1, 0, 7, 3, 0, 0},
-		{math.MaxInt64, 0, 3, 7, 0, 0},
-	}
-	for _, kase := range testCases {
-		t.Run(t.Name(), func(t *testing.T) {
-			buyingUnits, sellingUnits, e := convertToBuyingUnits(kase.sellingOfferAmount, kase.sellingUnitsNeeded, kase.pricen, kase.priced)
-			if !assert.Nil(t, e) {
-				return
-			}
-			assert.Equal(t, kase.wantBuyingUnits, buyingUnits)
-			assert.Equal(t, kase.wantSellingUnits, sellingUnits)
-		})
-	}
-}
-
 func TestWillAddOverflow(t *testing.T) {
 	testCases := []struct {
 		a                int64
